수정 split utils 위치 변경
