+++ src/test/java/calculator/SplitUtilsTest.java
@@ -1,6 +1,5 @@
-package calculator.utlis;
+package calculator;
 
-import calculator.utils.SplitUtils;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
